# Backend/bin/Release/net8.0/swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "bankApI, Version=1.0.0.0, Culture=neutral, PublicKeyToken=null",
    "version": "1.0"
  },
  "paths": {
...
    "/api/Client/UpdateClient": {
      "put": {
        "tags": [
          "Client"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/DUpdateClient"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/Client/AddClientNotification": {
      "post": {
        "tags": [
          "Client"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/DCNotifications"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/Client/AddGetHelpRequist": {
      "post": {
        "tags": [
          "Client"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/DCNotifications"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
    "/api/CNotifications/GetClientNotifications/{AccountId}": {
      "get": {
        "tags": [
          "CNotifications"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "AccountId",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/DCNotificationsGet"
              }
            }
          }
        }
      }
    },
    "/api/CNotifications/UpdateIsNotificationviewed/{Notificationid}": {
      "put": {
        "tags": [
          "CNotifications"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "Notificationid",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/Employee/AddNewEmployee": {
      "post": {
        "tags": [
          "Employee"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/DEmployeePerson"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/DGetEmployee"
            }
          }
        }
      }
    },
...
    "/api/Employee/UpdateEmployee": {
      "put": {
        "tags": [
          "Employee"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/DUpdateEmployee"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/Employee/FreezeUnfreezeEmployeeAccount": {
      "put": {
        "tags": [
          "Employee"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/DSetEmailState"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/Employee/SendMessageToEmployee": {
      "post": {
        "tags": [
          "Employee"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/DENotifications"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/ENotifications/GetEmployeeNotifications/{AccountId}": {
      "get": {
        "tags": [
          "ENotifications"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "AccountId",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/DCNotificationsGet"
              }
            }
          }
        }
      }
    },
    "/api/ENotifications/UpdateIsEmployeeNotificationviewed": {
      "put": {
        "tags": [
          "ENotifications"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/DSetIsNotificationV"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
    "/api/TransactionsHistory/GetTransactionHistoryById{AccountId}": {
      "get": {
        "tags": [
          "TransactionsHistory"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "AccountId",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/DGetTransactionsHistory"
              }
            }
          }
        }
      }
    },
    "/api/TransactionsHistory/TransferFund": {
      "put": {
        "tags": [
          "TransactionsHistory"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/DSetTransferFund"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "type": "string"
            }
          }
        }
      }
    }
  },
  "definitions": {
...
    "DCNotifications": {
      "type": "object",
      "properties": {
        "title": {
          "type": "string"
        },
        "body": {
          "type": "string"
        },
        "type": {
          "format": "int32",
          "type": "integer"
        },
        "accountId": {
          "type": "string"
        }
      },
      "additionalProperties": false
    },
    "DCNotificationsGet": {
      "type": "object",
      "properties": {
        "id": {
          "format": "int32",
          "type": "integer"
        },
        "notification": {
          "type": "string"
        },
        "title": {
          "type": "string"
        },
        "type": {
          "type": "string"
        },
        "date": {
          "format": "date",
          "type": "string"
        },
        "isviewed": {
          "type": "boolean"
        }
      },
      "additionalProperties": false
    },
...
    "DENotifications": {
      "type": "object",
      "properties": {
        "senderAccount": {
          "type": "string"
        },
        "title": {
          "type": "string"
        },
        "body": {
          "type": "string"
        },
        "type": {
          "format": "int32",
          "type": "integer"
        },
        "accountId": {
          "type": "string"
        }
      },
      "additionalProperties": false
    },
    "DEmployee": {
      "type": "object",
      "properties": {
        "salary": {
          "format": "double",
          "type": "number"
        },
        "typeId": {
          "format": "int32",
          "type": "integer"
        }
      },
      "additionalProperties": false
    },
    "DEmployeeAccount": {
      "type": "object",
      "properties": {
        "account": {
          "type": "string"
        },
        "password": {
          "type": "string"
        }
      },
      "additionalProperties": false
    },
    "DEmployeePerson": {
      "type": "object",
      "properties": {
        "person": {
          "$ref": "#/definitions/DPerson"
        },
        "employee": {
          "$ref": "#/definitions/DEmployee"
        },
        "employeeAccount": {
          "$ref": "#/definitions/DEmployeeAccount"
        }
      },
      "additionalProperties": false
    },
...
    "DGetEmployee": {
      "type": "object",
      "properties": {
        "id": {
          "format": "int32",
          "type": "integer"
        },
        "firstName": {
          "type": "string"
        },
        "lastName": {
          "type": "string"
        },
        "personalEmail": {
          "type": "string"
        },
        "birthDate": {
          "format": "date",
          "type": "string"
        },
        "address": {
          "type": "string"
        },
        "phoneNumber": {
          "type": "string"
        },
        "type": {
          "type": "string"
        },
        "roleType": {
          "type": "string"
        },
        "isActive": {
          "type": "boolean"
        },
        "salary": {
          "format": "double",
          "type": "number"
        }
      },
      "additionalProperties": false
    },
...
    "DGetTransactionsHistory": {
      "type": "object",
      "properties": {
        "n": {
          "format": "int32",
          "type": "integer"
        },
        "type": {
          "type": "string"
        },
        "amount": {
          "format": "double",
          "type": "number"
        },
        "date": {
          "format": "date",
          "type": "string"
        }
      },
      "additionalProperties": false
    },
...
    "DPerson": {
      "required": [
        "address",
        "birthDate",
        "email",
        "firstName",
        "lastName",
        "phoneNumber"
      ],
      "type": "object",
      "properties": {
        "firstName": {
          "maxLength": 100,
          "minLength": 1,
          "type": "string"
        },
        "lastName": {
          "maxLength": 100,
          "minLength": 1,
          "type": "string"
        },
        "birthDate": {
          "format": "date",
          "type": "string"
        },
        "email": {
          "minLength": 1,
          "type": "string"
        },
        "address": {
          "maxLength": 100,
          "minLength": 1,
          "type": "string"
        },
        "phoneNumber": {
          "minLength": 1,
          "type": "string"
        }
      },
      "additionalProperties": false
    },
...
    "DSetEmailState": {
      "type": "object",
      "properties": {
        "accountId": {
          "type": "string"
        },
        "state": {
          "type": "string"
        }
      },
      "additionalProperties": false
    },
    "DSetIsNotificationV": {
      "type": "object",
      "properties": {
        "id": {
          "format": "int32",
          "type": "integer"
        },
        "account": {
          "type": "string"
        }
      },
      "additionalProperties": false
    },
    "DSetTransferFund": {
      "type": "object",
      "properties": {
        "senderAccountId": {
          "type": "string"
        },
        "recieverAccountId": {
          "type": "string"
        },
        "amount": {
          "format": "double",
          "type": "number"
        },
        "description": {
          "type": "string"
        }
      },
      "additionalProperties": false
    },
    "DUpdateClient": {
      "type": "object",
      "properties": {
        "accountId": {
          "type": "string"
        },
        "firstName": {
          "type": "string"
        },
        "lastName": {
          "type": "string"
        },
        "birthDate": {
          "format": "date",
          "type": "string"
        },
        "email": {
          "type": "string"
        },
        "address": {
          "type": "string"
        },
        "phoneNumber": {
          "type": "string"
        }
      },
      "additionalProperties": false
    },
    "DUpdateEmployee": {
      "type": "object",
      "properties": {
        "accountId": {
          "type": "string"
        },
        "firstName": {
          "type": "string"
        },
        "lastName": {
          "type": "string"
        },
        "birthDate": {
          "format": "date",
          "type": "string"
        },
        "email": {
          "type": "string"
        },
        "address": {
          "type": "string"
        },
        "phoneNumber": {
          "type": "string"
        },
        "type": {
          "type": "string"
        }
      },
      "additionalProperties": false
    },
...
  }
}